pyproject.toml
[project]
name = "moviad-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "accelerate==0.23.0",
    "annotated-types==0.7.0",
    "av>=16.1.0",
    "calflops==0.3.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "docker-pycreds==0.4.0",
    "einops==0.8.1",
    "faiss-cpu==1.9.0.post1",
    "filelock==3.17.0",
    "fonttools==4.56.0",
    "fsspec==2025.2.0",
    "gitdb==4.0.12",
    "GitPython==3.1.44",
    "huggingface-hub==0.29.1",
    "idna==3.10",
    "imageio==2.37.0",
    "Jinja2==3.1.5",
    "joblib==1.4.2",
    "kiwisolver==1.4.8",
    "lazy_loader==0.4",
    "loguru==0.7.3",
    "MarkupSafe==3.0.2",
    "matplotlib>=3.10.8",
    "memory-profiler==0.61.0",
    "micromind==0.2.1",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numpy==2.2.3",
    "opencv-python==4.11.0.86",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "protobuf==5.29.3",
    "psutil==7.0.0",
    "ptflops>=0.7.5",
    "pydantic==2.10.6",
    "pydantic_core==2.27.2",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "PyYAML==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "safetensors==0.5.3",
    "scikit-image==0.25.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentry-sdk==2.22.0",
    "setproctitle==1.3.5",
    "setuptools==75.8.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sympy==1.13.3",
    "tabulate>=0.9.0",
    "threadpoolctl==3.5.0",
    "tifffile==2025.2.18",
    "timm>=1.0.24",
    "tokenizers==0.21.0",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tqdm==4.67.1",
    "transformers==4.49.0",
    "typing_extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "wandb==0.19.7",
    "wheel==0.45.1",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]


[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
